Refactor(image-gen): Centralize composition logic and fix bugs
This commit introduces a major refactoring of the image generation process to fix critical bugs related to state synchronization and data flow.

The root cause of the bugs (broken preview images and missing text fields) was that the automatic background generation process was not triggering the final composition of the images. The state was not being updated correctly and in the right sequence.

Changes:
1.  **New Utility:** A new reusable function, `composeSingleImage`, has been created in `src/utils/imageComposer.js`. It encapsulates all the logic for creating a single, complete composite image (background, brand elements, text, etc.).
2.  **Refactored `ImageGeneratorFrontendOnly.jsx`:** This component has been simplified. Its `generateImages` function now calls the new `composeSingleImage` utility, removing duplicated logic.
3.  **Refactored `HomePage.jsx`:** The `handleGenerateIAContent` function has been significantly improved. After the AI generates the text data, the component now does the following:
    -   Initializes all required state for the subsequent steps (`csvData`, `csvHeaders`, `fieldPositions`, etc.).
    -   If automatic image generation is enabled, it loops through each new post.
    -   In the loop, it first generates the unique background image for the post.
    -   It then immediately calls the new `composeSingleImage` utility to create the final composite image.
    -   It updates the application state with the complete image data (`url` and `blob`), ensuring the UI has the correct data when the user proceeds.
4.  **Feature Add:** As requested, the first successfully generated background image is now set as the global default `backgroundImage` for the campaign.
