+++ src/utils/imageComposer.js
@@ -1,85 +1,130 @@
-/**
- * Loads an image from a given URL.
- * @param {string} src - The source URL of the image.
- * @returns {Promise<HTMLImageElement>} A promise that resolves with the loaded image element.
- */
+import { containsHtml, renderHtmlToCanvas } from './htmlRenderer';
+
+// Helper functions moved from ImageGeneratorFrontendOnly.jsx and adapted for utility use
+
+const dataURLtoBlob = (dataurl) => {
+    if (!dataurl) return null;
+    const arr = dataurl.split(',');
+    if (arr.length < 2) return null;
+    const mimeMatch = arr[0].match(/:(.*?);/);
+    if (!mimeMatch) return null;
+    const mime = mimeMatch[1];
+    const bstr = atob(arr[1].split(';base64,').pop());
+    let n = bstr.length;
+    const u8arr = new Uint8Array(n);
+    while(n--){
+        u8arr[n] = bstr.charCodeAt(n);
+    }
+    return new Blob([u8arr], {type:mime});
+};
+
+const wrapTextInArea = (ctx, text, style, maxWidth, maxHeight) => {
+    if (!text) return [];
+    const fontSize = style.fontSize || 24;
+    const lineHeight = fontSize * (style.lineHeightMultiplier || 1.2);
+    const maxLines = Math.floor(maxHeight / lineHeight);
+    ctx.font = `${style.fontWeight || 'normal'} ${style.fontStyle || 'normal'} ${fontSize}px ${style.fontFamily || 'Arial'}`;
+    const words = text.toString().split(' ');
+    const lines = [];
+    let currentLine = words[0] || '';
+    for (let i = 1; i < words.length; i++) {
+        const word = words[i];
+        const testLine = currentLine + ' ' + word;
+        const metrics = ctx.measureText(testLine);
+        if (metrics.width > maxWidth && currentLine !== '') {
+            lines.push(currentLine);
+            if (lines.length >= maxLines) break;
+            currentLine = word;
+        } else {
+            currentLine = testLine;
+        }
+    }
+    if (lines.length < maxLines && currentLine) {
+        lines.push(currentLine);
+    }
+    return lines;
+};
+
+const applyTextEffects = (ctx, style) => {
+    ctx.fillStyle = style.color || '#000000';
+    ctx.font = `${style.fontWeight || 'normal'} ${style.fontStyle || 'normal'} ${style.fontSize || 24}px ${style.fontFamily || 'Arial'}`;
+    ctx.textAlign = style.textAlign || 'left';
+    ctx.textBaseline = 'top'; // Consistent baseline
+    if (style.textShadow) {
+        ctx.shadowColor = style.shadowColor || '#000000';
+        ctx.shadowBlur = style.shadowBlur || 4;
+        ctx.shadowOffsetX = style.shadowOffsetX || 2;
+        ctx.shadowOffsetY = style.shadowOffsetY || 2;
+    } else {
+        ctx.shadowColor = 'transparent';
+        ctx.shadowBlur = 0;
+        ctx.shadowOffsetX = 0;
+        ctx.shadowOffsetY = 0;
+    }
+    if (style.textStroke) {
+        ctx.strokeStyle = style.strokeColor || '#ffffff';
+        ctx.lineWidth = style.strokeWidth || 2;
+        ctx.lineJoin = 'round';
+        ctx.lineCap = 'round';
+    }
+};
+
+const drawTextWithEffects = async (ctx, text, x, y, style, maxWidth, maxHeight) => {
+    if (containsHtml(text)) {
+        await renderHtmlToCanvas(ctx, text, x, y, maxWidth, maxHeight, style);
+    } else {
+        if (style.textStroke) {
+            ctx.strokeText(text, x, y);
+        }
+        ctx.fillText(text, x, y);
+    }
+};
+
 const loadImage = (src) => {
   return new Promise((resolve, reject) => {
-    console.log(`[loadImage] Attempting to load image from: ${src}`);
     const img = new Image();
-    // For external images (like from Google Drive), this is necessary to avoid tainting the canvas.
     if (src.startsWith('http')) {
         img.crossOrigin = 'Anonymous';
     }
-    img.onload = () => {
-      console.log(`[loadImage] Successfully loaded image: ${src}`);
-      resolve(img);
-    };
-    img.onerror = (err) => {
-      console.error(`[loadImage] Failed to load image: ${src}`, err);
-      reject(new Error(`Failed to load image: ${src}`, { cause: err }));
-    };
+    img.onload = () => resolve(img);
+    img.onerror = (err) => reject(new Error(`Failed to load image: ${src}`, { cause: err }));
     img.src = src;
   });
 };
 
 /**
- * Composes a new image by combining a background, a logo, and a company image.
- * This version adds a colored rectangle at the bottom matching the company image's height.
+ * Composes a new image by applying filters and brand elements to a background.
  * @param {string} backgroundImageUrl - The URL or base64 string of the background image.
- * @param {string} logoUrl - The URL for the logo image (e.g., /LOGO.png).
- * @param {string} companyImageUrl - The URL for the company image (e.g., /EMPRESA.png).
- * @param {object} imageFilters - An object containing filter values.
- * @returns {Promise<string>} A promise that resolves with the data URL of the composed image.
+ * @param {object} imageFilters - An object containing filter values for the background.
+ * @param {Array<object>} brandElements - An array of brand element objects to overlay.
+ * @returns {Promise<HTMLCanvasElement>} A promise that resolves with the canvas containing the composed background.
  */
 export const composeImage = async (
   backgroundImageUrl,
   imageFilters = {},
   brandElements = []
 ) => {
-  console.log('[composeImage] Starting composition with:', {
-    backgroundImageUrl,
-    imageFilters,
-    brandElements
-  });
   try {
-    let backgroundSrc = backgroundImageUrl;
-    // A blob: URL is a valid source, so we should not treat it as a base64 string.
-    if (!backgroundImageUrl.startsWith('data:') && !backgroundImageUrl.startsWith('http') && !backgroundImageUrl.startsWith('blob:')) {
-      backgroundSrc = `data:image/png;base64,${backgroundImageUrl}`;
-    }
-
-    const bgImg = await loadImage(backgroundSrc);
-
+    const bgImg = await loadImage(backgroundImageUrl);
     const canvas = document.createElement('canvas');
     const ctx = canvas.getContext('2d');
-    const targetWidth = bgImg.width;
-    const targetHeight = bgImg.height;
-    canvas.width = targetWidth;
-    canvas.height = targetHeight;
+    canvas.width = bgImg.width;
+    canvas.height = bgImg.height;
 
     const { brightness = 100, contrast = 100, saturate = 100, blur = 0, opacity = 100 } = imageFilters;
-    const filterString = `brightness(${brightness}%) contrast(${contrast}%) saturate(${saturate}%) blur(${blur}px) opacity(${opacity}%)`;
-    ctx.filter = filterString;
-    ctx.drawImage(bgImg, 0, 0, targetWidth, targetHeight);
+    ctx.filter = `brightness(${brightness}%) contrast(${contrast}%) saturate(${saturate}%) blur(${blur}px) opacity(${opacity}%)`;
+    ctx.drawImage(bgImg, 0, 0);
     ctx.filter = 'none';
 
-    // Draw brand elements
     for (const element of brandElements) {
       if (!element.url) continue;
-
       try {
         const elementImg = await loadImage(element.url);
-
         ctx.save();
-
-        // Calculate pixel dimensions and positions
-        const elX = (element.x / 100) * targetWidth;
-        const elY = (element.y / 100) * targetHeight;
-        const elWidth = (element.width / 100) * targetWidth;
-        const elHeight = (element.height / 100) * targetHeight;
-
-        // Apply transformations (rotation)
+        const elX = (element.x / 100) * canvas.width;
+        const elY = (element.y / 100) * canvas.height;
+        const elWidth = (element.width / 100) * canvas.width;
+        const elHeight = (element.height / 100) * canvas.height;
         if (element.rotation) {
             const centerX = elX + elWidth / 2;
             const centerY = elY + elHeight / 2;
@@ -87,29 +132,126 @@ export const composeImage = async (
             ctx.rotate(element.rotation * Math.PI / 180);
             ctx.translate(-centerX, -centerY);
         }
-
-        // Apply individual filters
         if (element.filters) {
             const { brightness = 100, contrast = 100, saturate = 100, blur = 0, opacity = 100 } = element.filters;
             ctx.filter = `brightness(${brightness}%) contrast(${contrast}%) saturate(${saturate}%) blur(${blur}px) opacity(${opacity}%)`;
         }
-
         ctx.drawImage(elementImg, elX, elY, elWidth, elHeight);
-
-        ctx.restore(); // Restore context to remove filters and transformations for the next element
-
+        ctx.restore();
       } catch (error) {
         console.error(`[composeImage] Failed to load or draw brand element ${element.id}:`, error);
-        // Continue to the next element
       }
     }
-
-    const finalUrl = canvas.toDataURL('image/png');
-    console.log('[composeImage] Composition finished.');
-    return finalUrl;
+    return canvas;
   } catch (error) {
     console.error('Error composing image:', error);
-    // Propagate the error to be handled by the caller
     throw error;
   }
+};
+
+/**
+ * Creates a complete composite image with text and returns full imageData object.
+ * @param {object} params - The parameters for composition.
+ * @returns {Promise<object>} A promise that resolves with the final imageData object.
+ */
+export const composeSingleImage = async ({
+    record,
+    index,
+    itemBackgroundImage,
+    imageFilters,
+    brandElements,
+    fieldPositions,
+    fieldStyles
+}) => {
+    if (!itemBackgroundImage) {
+        throw new Error(`Background image is missing for record index ${index}.`);
+    }
+
+    // 1. Compose the background with filters and brand elements
+    const backgroundCanvas = await composeImage(itemBackgroundImage, imageFilters, brandElements);
+
+    // 2. Create a new canvas to draw the final image with text
+    const finalCanvas = document.createElement('canvas');
+    const ctx = finalCanvas.getContext('2d');
+    finalCanvas.width = backgroundCanvas.width;
+    finalCanvas.height = backgroundCanvas.height;
+    ctx.drawImage(backgroundCanvas, 0, 0);
+
+    // 3. Draw text fields onto the canvas
+    for (const field of Object.keys(record)) {
+        const position = fieldPositions[field];
+        const style = fieldStyles[field];
+        if (!position || !position.visible || !style) continue;
+        const text = record[field] || "";
+        if (!text) continue;
+
+        ctx.save();
+        const posPx = {
+            x: Math.round((position.x / 100) * finalCanvas.width),
+            y: Math.round((position.y / 100) * finalCanvas.height),
+            width: Math.round((position.width / 100) * finalCanvas.width),
+            height: Math.round((position.height / 100) * finalCanvas.height)
+        };
+
+        if (position.rotation) {
+            const centerX = posPx.x + posPx.width / 2;
+            const centerY = posPx.y + posPx.height / 2;
+            ctx.translate(centerX, centerY);
+            ctx.rotate(position.rotation * Math.PI / 180);
+            ctx.translate(-centerX, -centerY);
+        }
+
+        applyTextEffects(ctx, style);
+        const fixedPadding = 8;
+        const effectiveTextWidth = Math.max(0, posPx.width - (2 * fixedPadding));
+        const effectiveTextHeight = Math.max(0, posPx.height - (2 * fixedPadding));
+        const textContentStartX = posPx.x + fixedPadding;
+        const textContentStartY = posPx.y + fixedPadding;
+
+        const lines = wrapTextInArea(ctx, text, style, effectiveTextWidth, effectiveTextHeight);
+        const lineHeight = (style.fontSize || 24) * (style.lineHeightMultiplier || 1.2);
+
+        let currentLineRenderY = textContentStartY;
+        if (style.verticalAlign === 'middle') {
+            const totalTextBlockHeight = lines.length * lineHeight - (lines.length > 0 ? (lineHeight - (style.fontSize || 24)) : 0);
+            currentLineRenderY += (effectiveTextHeight - totalTextBlockHeight) / 2;
+        } else if (style.verticalAlign === 'bottom') {
+            const totalTextBlockHeight = lines.length * lineHeight - (lines.length > 0 ? (lineHeight - (style.fontSize || 24)) : 0);
+            currentLineRenderY += effectiveTextHeight - totalTextBlockHeight;
+        }
+
+        if (containsHtml(text)) {
+            await drawTextWithEffects(ctx, text, textContentStartX, textContentStartY, style, effectiveTextWidth, effectiveTextHeight);
+        } else {
+            for (const line of lines) {
+                let currentLineRenderX;
+                if (style.textAlign === 'center') {
+                    currentLineRenderX = textContentStartX + effectiveTextWidth / 2;
+                } else if (style.textAlign === 'right') {
+                    currentLineRenderX = textContentStartX + effectiveTextWidth;
+                } else {
+                    currentLineRenderX = textContentStartX;
+                }
+                const finalLineY = currentLineRenderY + (lines.indexOf(line) * lineHeight);
+                await drawTextWithEffects(ctx, line, currentLineRenderX, finalLineY, style, effectiveTextWidth, effectiveTextHeight);
+            }
+        }
+        ctx.restore();
+    }
+
+    // 4. Generate final data URL and blob
+    const dataUrl = finalCanvas.toDataURL('image/png', 1.0);
+    const blob = dataURLtoBlob(dataUrl);
+
+    // 5. Return the complete imageData object
+    return {
+        url: dataUrl,
+        dataUrl: dataUrl,
+        blob,
+        record,
+        index,
+        filename: `midiator_${String(index + 1).padStart(3, '0')}.png`,
+        backgroundImage: itemBackgroundImage,
+        // customFieldPositions and customFieldStyles are not handled here as this is for initial generation
+    };
 };
